Refine KDoc for methods and extensions
